pyproject.toml
[tool.poetry]
name = "alpha-get-values-b3"
version = "0.1.0"
description = ""
authors = ["moreirak14 <kaiquemoreiracloud@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dynaconf = "^3.1.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pylint = "^2.14.5"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
